api/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
google-cloud-bigquery==3.17.2
google-cloud-firestore==2.16.0
pydantic==2.6.0
